postapi/requirements.txt
Django>=3.2,<4.1
djangorestframework>=3.12
pytest>=6.0
pytest-django>=4.0